fix the redeem code through route
