+++ pages/claim/[type]/[code].tsx
@@ -13,6 +13,7 @@ import { UserPropsContext } from '../../../src/features/common/Layout/UserPropsC
 import { ErrorHandlingContext } from '../../../src/features/common/Layout/ErrorHandlingContext';
 import ShareOptions from '../../../src/features/common/ShareOptions/ShareOptions';
 import { styled } from '@mui/material';
+import CancelIcon from '../../../public/assets/images/icons/CancelIcon';
 
 const { useTranslation } = i18next;
 
@@ -87,10 +88,10 @@ function ClaimDonation({}: Props): ReactElement {
     }
   }, [router]);
 
-  async function validateCode(code: any, type: any) {
+  async function redeemCode(code: string) {
     setIsUploadingData(true);
     const submitData = {
-      type: type,
+      // type: type,
       code: code,
     };
     if (contextLoaded && user) {
@@ -124,8 +125,8 @@ function ClaimDonation({}: Props): ReactElement {
     // From here user can go back to home by clicking X
     if (contextLoaded && user) {
       // validate code
-      if (routerReady && code && type) {
-        validateCode(code, type);
+      if (routerReady && code) {
+        redeemCode(code);
       }
     }
 
@@ -144,34 +145,34 @@ function ClaimDonation({}: Props): ReactElement {
     }
   }, [contextLoaded, user, code]);
 
-  async function redeemCode(code: any, type: any) {
-    setIsUploadingData(true);
-    const submitData = {
-      type: type,
-      code: code,
-    };
-    if (contextLoaded && user) {
-      const userLang = localStorage.getItem('language') || 'en';
-      postAuthenticatedRequest(
-        `/api/v1.3/${userLang}/convertCode`,
-        submitData,
-        token,
-        handleError
-      ).then((res) => {
-        if (res.code === 401) {
-          setErrorMessage(res.message);
-          setIsUploadingData(false);
-        } else if (!res.response || res.response.status === 'error') {
-          setErrorMessage(res.errorText || t('me:wentWrong'));
-          setIsUploadingData(false);
-        } else if (res.response.status === 'success') {
-          setCodeRedeemed(true);
-          setIsUploadingData(false);
-          setCodeValidated(false);
-        }
-      });
-    }
-  }
+  // async function redeemCode(code: any, type: any) {
+  //   setIsUploadingData(true);
+  //   const submitData = {
+  //     type: type,
+  //     code: code,
+  //   };
+  //   if (contextLoaded && user) {
+  //     const userLang = localStorage.getItem('language') || 'en';
+  //     postAuthenticatedRequest(
+  //       `/api/v1.3/${userLang}/convertCode`,
+  //       submitData,
+  //       token,
+  //       handleError
+  //     ).then((res) => {
+  //       if (res.code === 401) {
+  //         setErrorMessage(res.message);
+  //         setIsUploadingData(false);
+  //       } else if (!res.response || res.response.status === 'error') {
+  //         setErrorMessage(res.errorText || t('me:wentWrong'));
+  //         setIsUploadingData(false);
+  //       } else if (res.response.status === 'success') {
+  //         setCodeRedeemed(true);
+  //         setIsUploadingData(false);
+  //         setCodeValidated(false);
+  //       }
+  //     });
+  //   }
+  // }
 
   const closeRedeem = () => {
     setCodeValidated(false);
@@ -287,61 +288,58 @@ function ClaimDonation({}: Props): ReactElement {
           <div className={styles.modal}>
             {codeValidated && validCodeData ? (
               <>
-                {errorMessage && (
-                  <span className={styles.formErrors}>{errorMessage}</span>
-                )}
+                <div style={{ position: 'relative', left: '108px' }}>
+                  <button className={styles.crossButton}>
+                    <CancelIcon />
+                  </button>
+                </div>
 
                 <div className={styles.codeTreeCount}>
                   {getFormattedNumber(
                     i18n.language,
-                    Number(validCodeData.treeCount)
+                    Number(validCodeData.units)
                   )}
                   <span>
                     {t('common:tree', {
-                      count: Number(validCodeData.treeCount),
+                      count: Number(validCodeData.units),
                     })}
                   </span>
                 </div>
 
-                {validCodeData.tpos?.length > 0 && (
-                  <div className={styles.plantedBy}>
-                    <span>{t('common:plantedBy')}</span>
-                    <p>{validCodeData.tpos[0].tpoName}</p>
-                  </div>
-                )}
-
-                <div
-                  onClick={() => redeemCode(code, type)}
-                  className="primaryButton"
-                  style={{ maxWidth: '200px', marginTop: '24px' }}
-                >
-                  {isUploadingData ? (
-                    <div className={styles.spinner}></div>
-                  ) : (
-                    t('redeem:addToMyTrees')
-                  )}
+                <div className={styles.codeTreeCount}>
+                  <span>{t('redeem:successfullyRedeemed')}</span>
                 </div>
               </>
             ) : (
               <>
-                <div>
-                  {t('redeem:validating')} {code}
+                <div style={{ position: 'relative', left: '108px' }}>
+                  <button className={styles.crossButton} onClick={closeRedeem}>
+                    <CancelIcon />
+                  </button>
                 </div>
-
-                {errorMessage && (
-                  <span className={styles.formErrors}>{errorMessage}</span>
+                {errorMessage ? (
+                  <div style={{ fontWeight: 'bold' }}>{code}</div>
+                ) : (
+                  <div style={{ fontWeight: 'bold' }}>
+                    {t('redeem:redeeming')} {code}
+                  </div>
                 )}
-                <div
-                  onClick={() => validateCode(code, type)}
-                  className="primaryButton"
-                  style={{ maxWidth: '200px', marginTop: '24px' }}
-                >
-                  {isUploadingData ? (
-                    <div className={styles.spinner}></div>
-                  ) : (
-                    t('redeem:validateCode')
-                  )}
-                </div>
+
+                {errorMessage ? (
+                  <span className={styles.formErrors}>{errorMessage}</span>
+                ) : (
+                  <div
+                    onClick={() => redeemCode(code)}
+                    className="primaryButton"
+                    style={{ maxWidth: '200px', marginTop: '24px' }}
+                  >
+                    {isUploadingData ? (
+                      <div className={styles.spinner}></div>
+                    ) : (
+                      t('redeem:redeemCode')
+                    )}
+                  </div>
+                )}
               </>
             )}
           </div>
