+++ src/features/user/Profile/styles/RedeemModal.module.scss
@@ -37,6 +37,9 @@
   top: -17px
 }
 
+.redeemCode{
+  margin-top: 24px;
+}
 
 
 .modalFinal {
@@ -104,7 +107,7 @@
 }
 
 .codeTreeCount {
-  font-size: fontXXXLarge;
+  font-size: 2rem;
   font-weight: bold;
   display: flex;
   flex-direction: row;
